separate controller read owners
