+++ crates/codestory-runtime/src/tests/activation_coverage_tests.rs
@@ -1,14 +1,17 @@
 use super::{
     hybrid_test_env, insert_semantic_fixture_nodes, test_index_publication, write_semantic_fixture,
 };
-use crate::{
-    AppController, CancellationToken, FileCoverageReason, IndexMode, IndexedFilesRequest,
-    PublicationTestAction, PublicationTestBoundary, Storage, Store,
-    arm_activation_search_before_revalidate_hook, arm_publication_test_fault,
-    load_persisted_search_state, read_search_generation_completion,
-    rebuild_search_state_from_storage, search_generation_completion_path,
-    search_index_path_for_publication, test_sidecar_runtime_from_env,
+use crate::search_publication::{
+    read_search_generation_completion, search_index_path_for_publication,
 };
+use crate::{
+    AppController, CancellationToken, PublicationTestAction, PublicationTestBoundary, Storage,
+    Store, arm_activation_search_before_revalidate_hook, arm_publication_test_fault,
+    load_persisted_search_state, rebuild_search_state_from_storage,
+    search_generation_completion_path, test_sidecar_runtime_from_env,
+};
+use codestory_contracts::api::{IndexMode, IndexedFilesRequest};
+use codestory_contracts::graph::FileCoverageReason;
 use std::fs;
 use tempfile::tempdir;
